intercept all objects
When the INTERCEPT_ALL_OBJS environment variable is set (to any value),
libsyscall_intercept attempts to patch all objects it finds. This is
not (yet) documented, as it is only meant to be used for a few testing
scenarios.
It also has caveats: the objects to be patched are filter based on their
names, thus if one libsyscall_intercept itself is not in loaded as a
library called libsyscall_intercept.so, it will try to intercept itself.
This can happen when it statically linked into some other object file.